CORE-5114 Refactor SandboxManagementService
